requirements.txt
streamlit==1.19.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
numpy==1.24.3
pymupdf==1.22.1
Pillow==9.5.0
huggingface_hub==0.14.0  # HuggingFace Hub for model downloads
altair==4.2.0
